generator: fix unnamed model library generation
Quick hack to give a name to models for which we have disabled the use
of the model name. The problem is that whilst we do not want the model
name for namespace and folder purposes, we still need it to determine
the library name. A proper solution will be backlogged.