generator: make things clearer around requiring stream manipulators
We had two different ways of saying the same thing: has boolean
properties and requires formatting. The former was used in the view
model whereas the latter was inclusion manager terminology. In both
cases what we are really saying is "does the type need to change the
stream by using stream manipulators?". Renamed both to make this a bit
clearer.