generator: add final to all classes which are not parents
At present we don't allow inheritance outside of the model because
this would cause problems with the way the code is generated - mainly
because we're being quite strict about not having non-abstract
non-leaf classes. So, just to make sure, all classes are now either
final (if they're leafs) or abstract (if they're not).