generator: propagate the nested type logic to view model
Unexpectedly we need to propagate type nesting all the way to the view
models and formatters. This is because the generators need to create
generating functions for each type argument in a generic
type (e.g. generate_vector, etc) so we cannot merge the types at the
view model level. This is unfortunate but unavoidable.