generator: implement support for keys per pod
In the past we had only a versioned/unversioned key pair for the
entire model. This has proved to be a problem. Instead we now are able
to toggle keys at the pod level by setting the stereotype
"cacheable". As part of this change, a lot of other fixes went in:

- remove database files. this was added when we were actively working
  on database support but since we stopped this story, it makes no
  sense to have half a feature in.
- add properties to empty classes that were being sanitized. sanitizer
  won't be able to validate classes with no properties