sanitizer: create a single registrar for all test models
The previous hack worked on lorenz by sheer luck it seems, as we had
multiple definitions for the same template function. As it happens the
linker was picking up the "right" one, e.g. boost, but it was luck of
the draw. The second version of the hack has a single registrar that
registers all models. This is of course more expensive but since its
test code its acceptable.