generator: one more attempt at fixing warnings
So the optimiser has own. Basically declaring something on an
anonymous namespace which does not get used within that translation
unit gives him a free licence to remove those symbols, and apparently
that is what its doing at least on OSX, triggering the lovely ranlib
warning. This commit moves back to the uniquely named dummy function
which we had originally and applies the same logic to the hack in file
outputter. Fingers crossed.