generator: change setters to a more usable format
We never intended to use pure immutability [1] in Dogen, but somehow
we got sucked into a lot of its problems. As discussed with IB, the
current state of getters and setters is not ideal: objects are
mutable, and designed to be used with const when immutable-like
behaviour is required - ideally, the Popsicle immutability Lippert
describes [1]. However, the get, change, set cycle done on object
setup is a pain in the backside without any obvious advantages.

This commit does three things:

- return a non-const reference to non-primitive data members so that
  one can change them in-place; this avoids the get, change, set
  cycle. Its a bit dirty conceptually but cleaner and more efficient
  code-wise. There is a lot of ceremony at the moment to change
  containers, etc.
- return a const reference to non-primitive data members. This weakens
  our move semantics, but I can't think of a better solution to avoid
  confusion between the above getter and this one. We need to see what
  the STL does for C++11.
- add move based setters. This is not the end-all solution for this,
  and we still need to revisit the Kolpackov series [2], but from
  where I stand, it seems like it doesn't make things worse.

[1] http://blogs.msdn.com/b/ericlippert/archive/2007/11/13/immutability-in-c-part-one-kinds-of-immutability.aspx
[2] http://www.codesynthesis.com/~boris/blog/2012/06/19/efficient-argument-passing-cxx11-part1/