dia to sml: refactor dia visitation logic with pre-processing
It was becoming increasingly obvious that the visitation for dia
objects was not the right strategy due to the amount of moving parts
it required at each usage point.

This approach instead uses a pre-processing of dia objects, putting
them in the right shape before anything is done to them.