cpp: remove standard based classes from model
Its just not possible to build the new object model incrementally as
the approach is too different. Instead, we'll leave cpp as is and
create a new model to experiment with. This commit largely restores
the model to its previous state (if with a somewhat bad layout).

New classes were preserved in the temporary model new_cpp_classes,
which is just a scrapbook.