agile: major refactoring of sprint's stories
The stories for this sprint were not really at the right level of
granularity - some were epic in nature rather than stories, others
were very jumbled up. Tidy-up the stories so that they make up a
logical (implementable) sequence.

Also, take into account the stitch refactoring in the redesign of the
formatters model.