yarn: resolve parent ids when collecting them
In the past we resolved the parent ids during the population phase of
the generalization expander. This does not work for trees with more than
2 levels because the order in which we are processing the elements is
not guaranteed. This means that if we process the child before we
process the parent, we will see the name prior to resolution. To solve
this problem we are now resolving as a first step.