yarn: process all types from all models in pipeline
With the introduction of cross-model inheritance, we can't ignore
external models any longer; if we do so, the generated code does not
take into account attributes from the external models. This fix changes
the yarn pipeline to process all models and fixes the resolver to take
this change into account. We also had to fix the references for some
models to include dependent models.