yarn: rebuild attribute names when obtained from concepts
Originally we took the name of the attribute directly as computed for
the concept, which makes sense. However, if we do this we cannot check
for duplicates by looking at all the IDs of the model elements. In
practice, it makes sense to think of attributes obtained from a
concept as if these attributes have moved in element space; they have
been cloned from the concept. This means that two objects that
obtained an attribute from a concept will have two distinct
attributes.