yarn.upsilon: process reference schemas correctly
Previously we were mangling the names of the types that belonged to
referenced schemas, placing them into the target schema. With this
commit, we now create a "mongrel" or "pre-merge-merged" yarn model
that is made up of the target plus all references. It seems to code
generate correctly - albeit it does generate types for all models,
target and references. In practice this is not an issue.