yarn: move enumeration and primitive resolution to resolver
We seem to incorrectly have placed this resolution in the
post-processing validator for some reason. However, given the current
state of the code, there is no good reason for this not to be done as
part of resolution (all the data is available in the indices) so move
it to the logical place - resolver.