yarn: fix a tricky validation problem with injected types
Generation of the ORM model northwind was failing with the new
injection validation. This is due to the fact we have injected
elements for both the forward declaration and the ODB options file,
sharing the same model name. Add large commentary explaining the
problem and remove duplicates checks for injected names.