cmake: clean-up targets for model generation
With the introduction of dogen.generation, we ran again into the
problem of generating dogen.generation.cpp when asking for
dogen.generation. The correct solution is still not clear but it was a
good opportunity to ensure JSON and Dia are in equal footing. This
also resolves the problem at hand.