generation.csharp: fix bad bug in csharp code generation
With one of the refactors of artefact properties, we somehow
fundamentally broke C# code generation. This went undetected because
the net result of the change (via a very convoluted path) was to
produce no diffs. This change reactivates C# code generation. A number
of fixes were made to ensure we see this through the diffs next time
round.