coding: generate a single visitor per model for a visitable type
We have been generating one visitor per internal module up to
now. This was not a problem because we had all model elements in the
same namespace. However, we recently changed that and created multiple
namespaces inside the meta-model and as a result we started creating
multiple visitors. This is not a desirable behaviour.

This commit buckets all leaves under the same bucket (instead of
creating a bucket per internal module), and as a result, creates a
single visitor per model. The downside is that cross model visitation
will now generate the visitor in the top-most namespace of that
model. This is not ideal, but its a lot less surprising than the
previous behaviour.