generation: remove checks for odb library
This check was actually not very useful because we do not know how the
user has detected CMake. By adding this check we widen the interface
between the dogen generated CMake and the manually crafted CMake for
no good reason. Instead, it is the responsibility of the calling code
to ensure that ODB was detected and configured correctly.