ctest: try to perform the update step on ci as well
At present we are skipping the CI step for version control
update. This is because it results in the CI provider doing a clone at
commit X and then we do a pull at (potentially) commit Y because a
commit may have happened in between. That was rather confusing so we
removed this step. We have now re-introduced the step as a "fake"
step: we reset head to the current commit, just as a way to tell ctest
about the commit we are building. This will hopefully link back to
CDash so that we can easily navigate to GitHub to get the commit
details as we could in the past.