github: comment out clang-cl build for now
At present, its not possible to mark a build as "allow failures". This
means that we cannot keep the experimental work in clang-cl as it is
causing all commits to have a red mark against them. This forces us to
check each and every build to see if any build, other than the clang-cl
builds, have failed. So for now we need to remove the clang-cl build
altogether, at least until GitHub actions support "allow failures". [1]

[1] https://github.com/actions/toolkit/issues/399